workflow pause: blockers ride the pause write, resume waits for every stored blocker
The integration-status reader moves below oauth_service so the workflow
layer imports it at module scope instead of deferring around the cycle.
